Add callbacks using references instead of smart pointers
Smart pointers are inappropriate for these callbacks, as they imply the recipient is responsible for the lifespan of the object (https://github.com/isocpp/CppCoreGuidelines/blob/master/CppCoreGuidelines.md#r30-take-smart-pointers-as-parameters-only-to-explicitly-express-lifetime-semantics). In most cases, the request / response objects are ephemeral and the caller should treat them as such.
Also, the constness of the pointers is prone to mistakes. For instance, a service callback can modify the request object since it is a `const shared_ptr<Request>`, not the `const shared_ptr<const Request>` that was probably intended.

Signed-off-by: Dan Rose <dan@digilabs.io>