More fixes for clang static analysis.
This is a smattering of changes to make clang static analysis
happier:

1.  It turns out that std::allocator::allocate does not return
nullptr on error (it throws an exception).  So there is no
reason to check for nullptr on the return, and a failing deleter
could indeed leak memory.  Instead catch any exceptions that
the deleter throws and manually deallocate and FAIL the test.
2.  It turns out that `get_parameter` can fail, so make sure
to initialize things we are checking against so that they will
fail if `get_parameter` fails for some reason.
3.  Many lambdas in test_function_traits.cpp are statically
asserted, but are never actually used at runtime.  Just
quiet down clang by using (void)name; on each of them.
4.  The C++ documentation that moved-from objects are in a
"valid but unspecified state".  To be truly cross-platform
(and please clang), don't do checks on the moved-from object.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>