Add in a fix for older compilers.
The addition of the NodeInterfaces class made it stop compiling
with older compilers (such as gcc 9.4.0 on Ubuntu 20.04).
The error has to do with calling the copy constructor on
rclcpp::Node, which is deleted.  Work around this by just
getting rid of the nullptr check; it isn't actually needed,
I don't think.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>