Add RCLCPP_*_FORMAT logging macros
Similar to the `RCLCPP_*_STREAM` family of macros, which enable the use
of C++'s stream-based output, these new macros allow users to
conveniently invoke the new C++20
[`std::format()`](https://en.cppreference.com/w/cpp/utility/format/format)
function for generating formatted text.

It should be noted that `rclcpp` (as of
[Iron](https://www.ros.org/reps/rep-2000.html#iron-irwini-may-2023-november-2024))
still has a minimum requirement of C++17, before `std::format` was
stabilized. Moreover, many compilers only recently implemented support
for this feature. (See the entry for P0645R10
[here](https://en.cppreference.com/w/cpp/compiler_support/20).)

To account for this, the relevant standard library header is only
included conditionally if it exists. This should guarantee that the
include doesn't introduce errors for compilers without or with partial
C++20 support, unless the user actually invokes the `RCLCPP_*_FORMAT`
macros.

These macros are not strictly necessary. As seen in the implementation,
one can mechanically replace any call to a `RCLCPP_*_FORMAT_*` macro
with the corresponding printf-style `RCLCPP_*_*` macro, passing in the
variadic arguments `"%s", std::format(...).c_str()`. However, this is
less readable and more error prone.

Signed-off-by: Nathan Wiebe Neufeldt <nwiebeneufeldt@clearpath.ai>