fix: Fixed possible memory violation after remove_node
Up until this change, after the call to remove node, entities of a
node might still be in the use by the executor. Therefore direct deletion
of the node after the call might result in a memory violation if an entity
would do a callback into the node itself. From now on, by default the
remove_node will be blocking until all entities are released by the executor,
which should be the expected behavior for most users.

Signed-off-by: Janosch Machowinski <J.Machowinski@cellumation.com>